The code has been modified
Added new component TrendingProducts, update code in netx components Header,LoginForm,CartPage,CertificatesPage,AboutSection,ProductDetails, and update code in files auth\operations.ts ,auth\slice.ts, servicesApi.ts
