+++ src/Interfaces/IloginError.ts
@@ -1,0 +1,3 @@
+export interface LoginError {
+  errorMessage: string;
+}
