--- src/pages/ProductDetails/productDeteils.css
@@ -1,3 +1,0 @@
-.swiper-slide {
-    transition: width 0.3s;
-  }
