+++ src/redux/auth/operations.ts
@@ -1,23 +1,20 @@
-import { KnownError } from "@redux/categories/operations";
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { authenticateUser } from "@services/servicesApi";
 import { AxiosError } from "axios";
+import { LoginError } from "Interfaces/IloginError";
 
-export const authUser = createAsyncThunk(
-  "auth/login",
-  async (
-    { email, password }: { email: string; password: string },
-    thunkAPI
-  ) => {
-    try {
-      const data = await authenticateUser(email, password);
-      return data;
-    } catch (err) {
-      const error = err as AxiosError<KnownError>;
-      if (!error.response) {
-        throw err;
-      }
-      return thunkAPI.rejectWithValue({ errorMessage: error.response.data });
-    }
+export const authUser = createAsyncThunk<
+  { token: string; name: string },
+  { email: string; password: string },
+  { rejectValue: LoginError }
+>("auth/login", async ({ email, password }, thunkAPI) => {
+  try {
+    const data = await authenticateUser(email, password);
+    return data;
+  } catch (err) {
+    const error = err as AxiosError<LoginError>;
+    return thunkAPI.rejectWithValue({
+      errorMessage: error.message || "Невідома помилка",
+    });
   }
-);
+});
